Moved items to new locations - `createNewPage` is now in its own file - `createNewBufferCursors`, `getNewline`, and `getLineStarts` now reside in `tree/tree.ts`
