+++ src/page/tree/tree.ts
@@ -2,7 +2,7 @@
  * Contains common items.
  */
 
-import { INode, IPageContent } from "../model";
+import { CharValues, IBufferCursor, INode, NEWLINE } from "../model";
 import { SENTINEL } from "../reducer";
 
 export interface INodePosition {
@@ -85,4 +85,77 @@ export function findNodeAtOffset(
   }
 
   return null!; // this will never be reached - it's just here to make the compiler happy.
+}
+
+/**
+ * Creates the start and end buffer cursors.
+ * @param startColumn The column of the start buffer.
+ * @param startLine The line of the start buffer.
+ * @param endColumn The column of the end buffer.
+ * @param endLine The line of the end buffer.
+ */
+export function createNewBufferCursors(
+  startColumn: number,
+  startLine: number,
+  endColumn: number,
+  endLine: number,
+): {
+  end: IBufferCursor;
+  start: IBufferCursor;
+} {
+  return {
+    start: {
+      column: startColumn,
+      line: startLine,
+    },
+    // tslint:disable-next-line:object-literal-sort-keys
+    end: {
+      column: endColumn,
+      line: endLine,
+    },
+  };
+}
+
+/**
+ * Checks the first 100 characters of a OneNote page to find what newline format is used. If it can't determine what
+ * format is used within the first 100 lines, it assumes that LF is used.
+ * @param content The HTML content of a OneNote page.
+ */
+export function getNewline(content: string): CharValues[] {
+  for (let i = 0; i < 100; i++) {
+    if (content.charCodeAt(i) === CharValues.LF) {
+      return NEWLINE.LF;
+    } else if (content.charCodeAt(i) === CharValues.CR) {
+      if (
+        i + 1 < content.length &&
+        content.charCodeAt(i + 1) === CharValues.LF
+      ) {
+        return NEWLINE.CRLF;
+      }
+    }
+  }
+  return NEWLINE.LF;
+}
+
+/**
+ * Gets an array of the indices of the line starts.
+ * @param content The HTML content of a OneNote page.
+ * @param newline The newline format for the OneNote page, as determined by the getNewline function.
+ */
+export function getLineStarts(
+  content: string,
+  newline: CharValues[],
+): number[] {
+  const lineStarts: number[] = [0];
+  for (let i = 0; i + newline.length - 1 <= content.length; i++) {
+    let match = true;
+    for (let j = 0; j < newline.length && match; j++) {
+      if (content.charCodeAt(i + j) !== newline[j]) {
+        match = false;
+      } else if (j === newline.length - 1) {
+        lineStarts.push(i + newline.length);
+      }
+    }
+  }
+  return lineStarts;
 }
